Add a new default target to build the JAR
The previous default target, "compile-test" failed:
https://github.com/andrewrjones/ant-task-sitemap/issues/1

This is because targets are defined whithout depends attribute,
while they expect other targets to run before.

This is fine, and I did that myself on several occasions, to avoid
running a whole build, which may take a long time, each time I want
to relaunch only one part of the build.

On the other hand, I find it useful to have a default target that
just works out of the box, by listing a sequence of targets in the
order in which they are expected to run.

I intended to build the JAR and run tests in this default target,
but the compilation of unit tests fails: org.apache.ant.tools.BuildFileTest
is missing. I will fix that first, and update the default target to
run the tests too.